+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Add the `cache.adapter.pdo_tag_aware` cache pool adapter
  * Add `framework.asset_mapper.importmap_integrity_algorithms` option to add integrity metadata to importmaps
+ * Add `framework.asset_mapper.minimum_release_age` option to delay JavaScript package updates until a version reaches a minimum age
  * Add `framework.mailer.smime_encrypter.certificates`, `on_missing_certificate` and `encrypt_for_sender` options
  * Add `framework.cache.default_provider` to configure `cache.app` with a DSN
  * Add `framework.messenger.reject_redelivered_messages` to allow disabling the `RejectRedeliveredMessageMiddleware`